backend/requirements.txt
# Core Web Framework
fastapi[all]
uvicorn

# AI and PDF Processing
openai
pypdf
python-multipart
aiofiles

# Data and Configuration
pydantic
python-dotenv
httpx

# Database
sqlmodel
asyncpg
aiosqlite

# Background Jobs
redis
taskiq
taskiq-redis

# Authentication
PyJWT

# AI Models
google-genai
langchain
